refactor(cartog): share one copy of the stable-version helpers
`parse_release_tag`, `is_stable_semver` and `compare_stable_versions` existed
as two byte-identical private copies: one in the lib's `auto_check`, one in the
bin's `commands::self_cmd::version`. The split was an artifact of the lib/bin
boundary, not a design choice.

Extract them into `cartog::semver`, following the `time_fmt` precedent — a
doc-hidden lib module that exists so the bin target can reach it without a
duplicate copy. `self_cmd::version` re-exports the three names so the
`use super::*` glob in every sibling submodule keeps resolving unchanged.

Move-only: no behaviour change. The existing `self_cmd/tests/version.rs` suite
already covers all three and now exercises the shared implementation, so no
new tests are added.
