+++ crates/cartog/src/semver.rs
@@ -1,0 +1,54 @@
+//! Stable-release version parsing and comparison.
+//!
+//! Shared by the background update check (`auto_check`) and the bin target's
+//! `cartog self` / `cartog doctor` commands, which all read the same GitHub
+//! releases payload and must agree on what counts as a stable release —
+//! three of them previously carried byte-identical private copies.
+
+/// Pull `tag_name` out of the GitHub release JSON, strip a leading `v`, and
+/// return `None` for any prerelease-shaped tag.
+///
+/// SemVer prerelease metadata is delimited by `-`, so any hyphen in the
+/// version (e.g. `0.15.0-rc.1`, `0.15.0-alpha`, `0.15.0-nightly.42`)
+/// disqualifies the tag.
+#[must_use]
+pub fn parse_release_tag(json: &str) -> Option<String> {
+    let parsed: serde_json::Value = serde_json::from_str(json).ok()?;
+    let tag = parsed.get("tag_name")?.as_str()?;
+    let trimmed = tag.strip_prefix('v').unwrap_or(tag);
+    if trimmed.contains('-') {
+        return None;
+    }
+    if !is_stable_semver(trimmed) {
+        return None;
+    }
+    Some(trimmed.to_string())
+}
+
+/// Quick guard: accept exactly three dot-separated non-empty numeric parts.
+#[must_use]
+pub fn is_stable_semver(s: &str) -> bool {
+    let parts: Vec<&str> = s.split('.').collect();
+    parts.len() == 3
+        && parts
+            .iter()
+            .all(|p| !p.is_empty() && p.bytes().all(|b| b.is_ascii_digit()))
+}
+
+/// Lexicographic compare on `(major, minor, patch)`.
+///
+/// Both inputs are expected to be stable `MAJOR.MINOR.PATCH` triples — any
+/// non-numeric component is treated as `0`, so the function never panics on
+/// weird input but degrades gracefully.
+#[must_use]
+pub fn compare_stable_versions(a: &str, b: &str) -> std::cmp::Ordering {
+    let parse = |s: &str| -> [u64; 3] {
+        let mut parts = s.split('.').map(|p| p.parse::<u64>().unwrap_or(0));
+        [
+            parts.next().unwrap_or(0),
+            parts.next().unwrap_or(0),
+            parts.next().unwrap_or(0),
+        ]
+    };
+    parse(a).cmp(&parse(b))
+}
